refactor: split vector search tests
